incorporate pr comments
